Update 11thMarch.py
Write a python code to print the position or index of a given string(taken as an input from user) from a given list of strings.

The program will take an input from user in the form of a string and will pass the string as argument to a function. The function will take the string as argument and return position(or index) of the list if the passed string is present in the list, else it'll return “String not found”. If the passed string is present at multiple indices, in that case the function should only return the first index of occurrence.